Rename to alert_error
