+++ lib/domain/premium_introduction/components/premium_introduction_footer_store.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:pilll/analytics.dart';
 import 'package:pilll/domain/premium_introduction/util/map_to_error.dart';
-import 'package:pilll/entity/user_error.dart';
+import 'package:pilll/entity/alert_error.dart';
 import 'package:pilll/error_log.dart';
 import 'package:pilll/service/purchase.dart';
 import 'package:purchases_flutter/purchases_flutter.dart';
