+++ lib/native/health_care.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:pilll/entity/menstruation.codegen.dart';
-import 'package:pilll/entity/user_error.dart';
+import 'package:pilll/entity/alert_error.dart';
 import 'package:pilll/native/channel.dart';
 
 Future<bool> isHealthDataAvailable() async {
